requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
frozenlist==1.4.0
google==3.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.1.0
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2